# spec/api-openapi.json
{
  "components": {
    "schemas": {
      "AgentRunOut": {
        "properties": {
          "cache_hit": {
            "title": "Cache Hit",
            "type": "boolean"
          },
          "cost_est": {
            "title": "Cost Est",
            "type": "number"
          },
          "degraded": {
            "title": "Degraded",
            "type": "boolean"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "input_tokens": {
            "title": "Input Tokens",
            "type": "integer"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "node": {
            "title": "Node",
            "type": "string"
          },
          "output_tokens": {
            "title": "Output Tokens",
            "type": "integer"
          },
          "retry_count": {
            "title": "Retry Count",
            "type": "integer"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          }
        },
        "required": [
          "node",
          "input_tokens",
          "output_tokens",
          "cache_hit",
          "duration_ms",
          "cost_est",
          "retry_count",
          "degraded"
        ],
        "title": "AgentRunOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlanConfirmBody": {
        "properties": {
          "expected_attempt": {
            "title": "Expected Attempt",
            "type": "integer"
          },
          "plan": {
            "additionalProperties": true,
            "title": "Plan",
            "type": "object"
          }
        },
        "required": [
          "plan",
          "expected_attempt"
        ],
        "title": "PlanConfirmBody",
        "type": "object"
      },
...
      "TaskControlOut": {
        "properties": {
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "status"
        ],
        "title": "TaskControlOut",
        "type": "object"
      },
      "TaskDetailOut": {
        "properties": {
          "batch_task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Task Id"
          },
          "chapter_seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Seq"
          },
          "cost_total": {
            "default": 0.0,
            "title": "Cost Total",
            "type": "number"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskProgressOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "retry_count": {
            "title": "Retry Count",
            "type": "integer"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/AgentRunOut"
            },
            "title": "Runs",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "task_type": {
            "title": "Task Type",
            "type": "string"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          }
        },
        "required": [
          "task_id",
          "task_type",
          "status",
          "retry_count"
        ],
        "title": "TaskDetailOut",
        "type": "object"
      },
      "TaskProgressOut": {
        "properties": {
          "current": {
            "title": "Current",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "current",
          "total"
        ],
        "title": "TaskProgressOut",
        "type": "object"
      },
      "TaskSummaryOut": {
        "description": "任务历史列表项（§阶段 4 任务视图）：轻量摘要不含 runs，点开再拉详情。\n\n批次进度为派生量（同 TaskDetailOut.progress 口径）：current = 该批 persist 节点\n去重章数，total = payload.size。观测表（tasks 无 RLS），new_session 普通连接可查。",
        "properties": {
          "batch_current": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Current"
          },
          "batch_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Size"
          },
          "chapter_seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Seq"
          },
          "cost_total": {
            "default": 0.0,
            "title": "Cost Total",
            "type": "number"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "task_type": {
            "title": "Task Type",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "task_type",
          "status"
        ],
        "title": "TaskSummaryOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Myink 内部 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/v1/projects/{project_id}/tasks": {
      "get": {
        "description": "项目任务历史（倒序）：全项目最近 50 条；按章节查询时扫描完整历史。\n\n轻量摘要不含 runs（重量留给 GET /tasks/{id}）；批次进度派生（persist 去重章数）。\nrequire_owner 归属断言由依赖挂载（§14.1 ③）；tasks 无 RLS 观测表，new_session 可查。\n\nchapter_seq 非空（§右栏按章过滤）：只返回覆盖该章的任务——单章按 chapter_seq 精确\n匹配、批次按 payload.start..start+size 范围覆盖；cost_total 随之收窄为该章 agent_runs\n切片（批次 = {batch_id}:ch{seq} 行，单章 = 全任务；批次 book 级 run 如 batch_plan 不计入章成本）。",
        "operationId": "list_project_tasks_internal_v1_projects__project_id__tasks_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "chapter_seq",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Chapter Seq"
            }
          },
          {
            "in": "header",
            "name": "X-Myink-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Myink-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskSummaryOut"
                  },
                  "title": "Response List Project Tasks Internal V1 Projects  Project Id  Tasks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Project Tasks",
        "tags": [
          "tasks"
        ]
      }
    },
...
    "/internal/v1/tasks/{task_id}": {
      "get": {
        "operationId": "get_task_internal_v1_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/internal/v1/tasks/{task_id}/cancel": {
      "post": {
        "operationId": "cancel_task_internal_v1_tasks__task_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskControlOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/internal/v1/tasks/{task_id}/pause": {
      "post": {
        "operationId": "pause_task_internal_v1_tasks__task_id__pause_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskControlOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pause Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/internal/v1/tasks/{task_id}/plan/confirm": {
      "post": {
        "description": "确认手动模式计划，用同一任务 ID 从 plan_gate 断点继续。",
        "operationId": "confirm_task_plan_internal_v1_tasks__task_id__plan_confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanConfirmBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskControlOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Task Plan",
        "tags": [
          "tasks"
        ]
      }
    },
    "/internal/v1/tasks/{task_id}/resume": {
      "post": {
        "operationId": "resume_task_internal_v1_tasks__task_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskControlOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume Task",
        "tags": [
          "tasks"
        ]
      }
    },
...
  }
}